Fix major GLTF material indexing bug; lots of deferred shading progress
Resolved a major GLTF rendering bug where the material buffer was not populated properly and thus lead to incorrect materials being applied to models.
Additionally, I've finally gotten the end to end setup for deferredShading implemented, but there are still bugs wrt material application, lighting and optimized barriers in the rendering loop that need to be addressed.